rename the challenge file and deleting unwanted files
